Removed static from functions. Added main class.
Refactored the code to use classes in a less silly way. Each class now has a definitive purpose, and can be instantiated. Grammar class still has a static function, but that was intentional because it makes some other code simpler.

Removed the folder "stanford-parser-full-2017-06-09", and moved the 2 files that were in it that we're using into packages. This should make the file hierarchy simpler and smaller. There is still a LOT we need to do to reduce the project's size, but this was just the first step.

Also cleaned up some portions of the code and made sure it was formatted to normal C# standards (fine, I'll use your dumb '{' on next line thing).